% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bees}
\alias{bees}
\title{bees: a data file bee activity by gender}
\format{
a dataframe containing experimental results exploring how temperature affects activity in male and female bees. Each row is a result from a single bee
\describe{
 \item{Temp}{Temperature (degrees celsius)}
 \item{Active}{A categorial factor  Yes / No}
 \item{Gender}{Bee gender}
}
}
\source{
unknown
}
\usage{
bees
}
\description{
bees: a data file bee activity by gender
}
\keyword{datasets}
