% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bloodpressure}
\alias{bloodpressure}
\title{bloodpressure: A data file on systolic blood pressure with diet change}
\format{
a dataframe containing individual systolic blood pressure measurements before and after changing to a vegatarian diet.
\describe{
 \item{Subject}{Numeric code for each individual in the study}
 \item{Before}{Measured systolic blood pressure prior to diet change"}
 \item{After}{Meausred systolic blood pressure after diet change}
}
}
\source{
Rosner 1995. Fundamentals of Biostatistics. Duxbury Press. 4th Edition
}
\usage{
bloodpressure
}
\description{
bloodpressure: A data file on systolic blood pressure with diet change
}
\keyword{datasets}
