% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{CarsMPG}
\alias{CarsMPG}
\title{CarsMPG: a data file of car fuel efficiency}
\format{
a dataframe data on fuel consumption , weight, model year, and price for 245 cars produced in the United States between 1970 and 1980
\describe{
 \item{mg}{fuel consumption (miles per gallon)}
 \item{weight}{car weight (points)}
 \item{weight}{car weight (points)}
 \item{price}{retail price (U.S. dollars)}
 \item{model.year}{year that the car model was produced, 1970-1982, truncated to two digits i.e 70 = 1970}
}
}
\usage{
CarsMPG
}
\description{
CarsMPG: a data file of car fuel efficiency
}
\keyword{datasets}
