% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{childPoverty}
\alias{childPoverty}
\title{childPoverty: A data file on Washington State child poverty rates}
\format{
a dataframe containing number of children in poverty, by city and by race.  Data are five year averages ending in 2018
\describe{
 \item{city_name}{Name of city / town}
 \item{group_name}{Racial group}
 \item{number}{total number of children in poverty}
 \item{total_population}{total population size for each race in each city}
 \item{percent}{Percentage of residents of each city and racial group experiencing childhood poverty}
}
}
\source{
https://www.cityhealthdashboard.com/data-access
}
\usage{
childPoverty
}
\description{
childPoverty: A data file on Washington State child poverty rates
}
\keyword{datasets}
