% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{lead}
\alias{lead}
\title{lead: A data file on lead levels and neurological effects in children}
\format{

}
\source{
Landrigan  et al. (1975) "Neuropsychological dysfunction in children with chronic low-level lead absorption".  The Lancet,  March 29, pp. 708--712
}
\usage{
lead
}
\description{
lead: A data file on lead levels and neurological effects in children
}
\keyword{datasets}
