% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{LifeSpan}
\alias{LifeSpan}
\title{LifeSpan: A data file on Life Expectancy by U.S. City}
\format{
a dataframe containing city level estimates of average life expectancy, 2010 - 2015
\describe{
 \item{state}{Two letter state abbreviation}
 \item{city_name}{City}
 \item{life_expectance}{Expected number of years that an individual will life after birth}
 \item{data_yr_type}{Detailed information on method and year range for life expectancy estimate}
}
}
\source{
Department of Population Health, NYU Langone Health. City Health Dashboard. https://www.cityhealthdashboard.com/. Accessed August 11, 2020.
}
\usage{
LifeSpan
}
\description{
LifeSpan: A data file on Life Expectancy by U.S. City
}
\keyword{datasets}
