% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{NBP}
\alias{NBP}
\title{NBP: a data file of the number of bird observers and numbers of robins seen in Discovery Partk}
\format{
a dataframe showing bird observer data from the National Audubon Society's Nieghborhood bird project.  Volunteers with this project conduct monthly surveys of natural areas within the city. All birds that are seen or heard within 50 meters of a monitoring station are recorded.
\describe{
 \item{Observers}{Number of volunteers observing birds}
 \item{Tally}{Number of robins seen}
}
}
\usage{
NBP
}
\description{
NBP: a data file of the number of bird observers and numbers of robins seen in Discovery Partk
}
\keyword{datasets}
