% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{NusaPenida}
\alias{NusaPenida}
\title{NusaPenida: a data file of GPS (geographical information system) bird locations}
\format{
a dataframe containing GPS location ouputs of an invasive bird species tagged on the island of Nusa Penida, Indonesia.
\describe{
 \item{Year}{Year of observation}
 \item{Day}{Julian day}
 \item{Hour}{Hour in 24-Hour notation}
 \item{Minutes}{Minutes after the hour}
 \item{AirTemp}{Air Temperature (degrees celsius)}
 \item{Latitude}{Latitude, in degrees S}
 \item{Longitude}{Longitude in degrees E}
 \item{HDOP}{Horizontal Dilution of Precision, which is an indicator of the accuracy for each recording (higher values indicate lower accuracy)}
}
}
\source{
Abrantes, A.A. 2018. 'Avifaunal Biodiversity and Land Use on Indonesia's Palau Penida Archipelago'. MS Thesis. State University of New York at Fredonia. Fredonia, NY
}
\usage{
NusaPenida
}
\description{
NusaPenida: a data file of GPS (geographical information system) bird locations
}
\keyword{datasets}
