% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{PMdata}
\alias{PMdata}
\title{PMdata: a data file  average daily concentration of fine particulate matter (PM)}
\format{
a dataframe average daily concentration of fine particulate matter measured from an air quality monitoring station in Seattle, Washington, during summer 2017 and summer 2018.  Each row is a day.
\describe{
 \item{PM2017}{fine PM (micro grams per cubic meter) in summer 2017}
 \item{PM2018}{fine PM (micro grams per cubic meter) in summer 2018}
}
}
\usage{
PMdata
}
\description{
PMdata: a data file  average daily concentration of fine particulate matter (PM)
}
\keyword{datasets}
