% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QSCI381-package.R
\docType{package}
\name{QSCI381-package}
\alias{QSCI381}
\alias{QSCI381-package}
\title{QSCI381: Data used in QSCI 381, University of Washington}
\description{
This package uploads data sets used by QSCI instructors.  Some instructors will also teach you how to import data directly into R.  Data in this package came from a variety of sources, including the R packages "openintro", "Stat2Data", and "moderndive"
}
\details{
Contains datafiles that you will use in QSCI 381 at the University of Washington.
}
\keyword{internal}
