% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{WAtree}
\alias{WAtree}
\title{WAtree: a data file of tree widths}
\format{
a dataframe containing 1,100 measurements of tree diameters for two species.  Each row is a single tree / species combination
\describe{
 \item{dfir}{diameter at breast hieght of Douglas Fir}
 \item{pine}{diameter at breast hieght of Ponderosa Pine}
}
}
\usage{
WAtree
}
\description{
WAtree: a data file of tree widths
}
\keyword{datasets}
